Define implementations in StructAccess
Slightly improve compilation time and decrease the amount of code
present in modules which use StructAccess by calling out to functions
defined in StructAccess.

No longer implement the `Access.get/2` callback as this was removed as a
requirement in Elixir 1.7.0.

Make the StructAccess `Access` implementations overridable.

Bump the required Elixir version to "~> 1.7"

Bump the version to 1.1.0